+++ PSModuleDevelopment/functions/templating/Invoke-PSMDTemplate.ps1
@@ -112,7 +112,7 @@
 		$NoFolder,
 		
 		[hashtable]
-		$Parameters,
+		$Parameters = @{ },
 		
 		[switch]
 		$Raw,
@@ -129,8 +129,6 @@
 	
 	begin
 	{
-		Write-PSFMessage -Level InternalComment -Message "Bound parameters: $($PSBoundParameters.Keys -join ", ")" -Tag 'debug', 'start', 'param'
-		
 		#region Validate output path
 		try
 		{