+++ templates/PSFTests/general/strings.Tests.ps1
@@ -12,6 +12,7 @@ $exceptions = & "$PSScriptRoot\strings.Exceptions.ps1"
 
 Describe "Testing localization strings" {
     foreach ($stringEntry in $stringsResults) {
+        if ($stringEntry.String -eq "key") { continue } # Skipping the template default entry
         It "Should be used & have text: $($stringEntry.String)" {
             if ($exceptions.LegalSurplus -notcontains $stringEntry.String) {
                 $stringEntry.Surplus | Should -BeFalse